feat: reimplement `Dictionary` with graphs
Reimplement the `Dictionary` using `petgraph::Graph` as internal data
structure. This not only makes the `Dictionary` implementation simpler,
it also adds the benefit that `Dictionary` implements `Send` and `Sync`.
Of course, the general rules apply here as well, `Send` and `Sync` are
only implemented if `Key` and `Value` implement them as well.